requirements.txt
opencv-python==4.10.0.84
PyOpenGL==3.1.7
numpy==1.26.4
matplotlib
tqdm